Move and rename all cli.cmd_xxx into cli.cmd.xxx.
Why simulate what can just as well be a package name using a file name
prefix?
