+++ pkg/client/cli/cmd/connect.go
@@ -1,0 +1,28 @@
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+
+	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/ann"
+	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/connect"
+	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/daemon"
+)
+
+func connectCmd() *cobra.Command {
+	var request *daemon.Request
+
+	cmd := &cobra.Command{
+		Use:   "connect [flags] [-- <command to run while connected>]",
+		Args:  cobra.ArbitraryArgs,
+		Short: "Connect to a cluster",
+		Annotations: map[string]string{
+			ann.Session: ann.Required,
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			request.CommitFlags(cmd)
+			return connect.RunConnect(cmd, args)
+		},
+	}
+	request = daemon.InitRequest(cmd)
+	return cmd
+}
