+++ pkg/client/cli/cmd/quit.go
@@ -1,38 +1,20 @@
-package connect
+package cmd
 
 import (
 	"fmt"
 	"os"
 
 	"github.com/spf13/cobra"
-	empty "google.golang.org/protobuf/types/known/emptypb"
+	"google.golang.org/protobuf/types/known/emptypb"
 
-	rpc "github.com/telepresenceio/telepresence/rpc/v2/daemon"
+	daemon2 "github.com/telepresenceio/telepresence/rpc/v2/daemon"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/ann"
+	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/connect"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/daemon"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/socket"
 )
 
-func Command() *cobra.Command {
-	var request *daemon.Request
-
-	cmd := &cobra.Command{
-		Use:   "connect [flags] [-- <command to run while connected>]",
-		Args:  cobra.ArbitraryArgs,
-		Short: "Connect to a cluster",
-		Annotations: map[string]string{
-			ann.Session: ann.Required,
-		},
-		RunE: func(cmd *cobra.Command, args []string) error {
-			request.CommitFlags(cmd)
-			return RunConnect(cmd, args)
-		},
-	}
-	request = daemon.InitRequest(cmd)
-	return cmd
-}
-
-func QuitCommand() *cobra.Command {
+func quit() *cobra.Command {
 	quitDaemons := false
 	quitRootDaemon := false
 	quitUserDaemon := false
@@ -43,7 +25,7 @@ func QuitCommand() *cobra.Command {
 		Short:       "Tell telepresence daemon to quit",
 		Annotations: map[string]string{ann.UserDaemon: ann.Optional},
 		RunE: func(cmd *cobra.Command, _ []string) error {
-			if err := InitCommand(cmd); err != nil {
+			if err := connect.InitCommand(cmd); err != nil {
 				return err
 			}
 			if quitUserDaemon {
@@ -59,10 +41,10 @@ func QuitCommand() *cobra.Command {
 				// User daemon isn't running. If the root daemon is running, we must
 				// kill it from here.
 				if conn, err := socket.Dial(ctx, socket.DaemonName); err == nil {
-					_, _ = rpc.NewDaemonClient(conn).Quit(ctx, &empty.Empty{})
+					_, _ = daemon2.NewDaemonClient(conn).Quit(ctx, &emptypb.Empty{})
 				}
 			}
-			return Disconnect(cmd.Context(), quitDaemons)
+			return connect.Disconnect(cmd.Context(), quitDaemons)
 		},
 	}
 	flags := cmd.Flags()
