+++ pkg/client/cli/cmd/testdata/license
