+++ pkg/client/cli/cmd/testdata/testLogDir/cli.log
