+++ pkg/client/cli/cmd/testdata/testLogDir/connector-20210916T130347.log
