+++ pkg/client/cli/cmd/testdata/testLogDir/connector-20210916T130624.log
