+++ pkg/client/cli/cmd/testdata/testLogDir/connector.log
