+++ pkg/client/cli/cmd/testdata/testLogDir/daemon-20210916T130318.log
