+++ pkg/client/cli/cmd/testdata/testLogDir/daemon-20210916T130402.log
