+++ pkg/client/cli/cmd/testdata/testLogDir/daemon.log
