+++ pkg/client/cli/cmd/testdata/testLogDir/echo-auto-inject-6496f77cbd-n86nc
