+++ pkg/client/cli/cmd/testdata/testLogDir/traffic-manager-5c69859f94-g4ntj
