+++ pkg/client/cli/cmd/testdata/zipDir/diff_name.log
