+++ pkg/client/cli/cmd/testdata/zipDir/file1.log
