+++ pkg/client/cli/cmd/testdata/zipDir/file2.log
